+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2026-01-26
+
+### Added
+- Memo Replies: Support for a second level of memos (threaded conversations)
+- Reply button in memo action tooltip
+
+### Changed
+- Renamed "Task Indentation" setting to "Markdown Indentation" (applies to both tasks and memos)
+
 ## [1.1.3] - 2026-01-26
 
 ### Added